delete custom BOM code due to octopart API shutdown
The octopart v4 API has a *monthly* limit of 500 parts/month,
which we got close to exceeding within just a few hours.

Clearly, our usage is not sustainable on that API,
so just link to the shared BOMs on the octopart BOM tool instead.
This gives users a way to see the current cost, too
(just not on our own site).